requirements.txt
asgiref==3.5.0
black==21.12b0
click==8.0.1
colorama==0.4.4
dj-rest-auth==2.2.4
Django==4.0.3
django-filter==21.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
drf-nested-routers==0.93.4
mypy-extensions==0.4.3
mysqlclient==2.1.0
pathspec==0.9.0
Pillow==9.1.0
platformdirs==2.5.1
PyJWT==2.3.0
python-dotenv==0.20.0
pytz==2022.1
sqlparse==0.4.2
tomli==1.2.3
typing_extensions==4.1.1
tzdata==2022.1
